Specify how I'd like this buildpack should work. (README-driven development in this case)
@jasonkarns & @theotherzach: feedback on whether this
configuration would be robust enough? What I like 
is that it's not Rails-specific, that it supports
multiple apps, and that it shouldn't be rocket science
to implement given that we already have the internal
behavior working as a shell script in:
https://github.com/testdouble/heroku-buildpack-lineman

If either of you have the gumption to tackle this
on your own, please go for it! Otherwise I'll try 
to tackle this soonish.